📝 ✔️ Struct, Test: Move special extensions back to mother class.
