+++ Common.BasicHelper.Test/Core/Shell/CommandsExecutor_Tests.cs
@@ -3,5 +3,15 @@
 [TestClass]
 public class CommandsExecutor_Tests
 {
+    [TestMethod]
+    public void Test_ExecuteAsCommand()
+    {
+        Console.WriteLine("help".ExecuteAsCommand());
+    }
 
+    [TestMethod]
+    public async Task Test_ExecuteAsCommandAsync()
+    {
+        Console.WriteLine(await "help".ExecuteAsCommandAsync());
+    }
 }
