+++ Common.BasicHelper.Test/Utils/Extensions/StringHelper_Tests.cs
@@ -38,16 +38,4 @@ public class StringHelper_Tests
         Assert.IsTrue(c.IsNullOrWhiteSpace());
         Assert.IsTrue(d.IsNullOrWhiteSpace());
     }
-
-    [TestMethod]
-    public void Test_ExecuteAsCommand()
-    {
-        Console.WriteLine("help".ExecuteAsCommand());
-    }
-
-    [TestMethod]
-    public async Task Test_ExecuteAsCommandAsync()
-    {
-        Console.WriteLine(await "help".ExecuteAsCommandAsync());
-    }
 }
