+++ Common.BasicHelper/Core/Shell/CommandsExecutor.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Diagnostics;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -108,4 +107,57 @@ public static class CommandsExecutor
         return output;
     }
 
+}
+
+public static class CommandsExecutorExtensions
+{
+
+    /// <summary>
+    /// 将当前字符串作为命令执行, 返回命令执行的输出
+    /// </summary>
+    /// <param name="command">命令</param>
+    /// <param name="args">命令参数</param>
+    /// <param name="action">针对启动信息的行动</param>
+    /// <returns>执行的输出</returns>
+    public static string ExecuteAsCommand
+    (
+        this string command,
+        string? args = null,
+        bool findInPath = true,
+        Action<ProcessStartInfo>? action = null
+    )
+        => CommandsExecutor.GetExecutionResult
+        (
+            command,
+            args ?? "",
+            findInPath,
+            action
+        );
+
+
+    /// <summary>
+    /// 将当前字符串作为命令执行, 异步获取命令执行输出
+    /// </summary>
+    /// <param name="command">命令</param>
+    /// <param name="args">参数</param>
+    /// <param name="findInPath">是否在 Path 中寻找</param>
+    /// <param name="action">针对启动信息的动作</param>
+    /// <param name="token">取消口令</param>
+    /// <returns>命令执行输出</returns>
+    public static Task<string> ExecuteAsCommandAsync
+    (
+        this string command,
+        string? args = null,
+        bool findInPath = true,
+        Action<ProcessStartInfo>? action = null,
+        CancellationToken? token = default
+    )
+        => CommandsExecutor.GetExecutionResultAsync
+        (
+            command,
+            args ?? "",
+            findInPath,
+            action,
+            token
+        );
 }
